ovirt: Ask for the CA pem content ovirt API
Some components of the cluster are interacting with oVirt API but don't
have way to verify it's certificate.

To support that the installer will ask for the CA bundle, and it
will be kept inside ovirt-credentials secret.

Once in ovirt-credentials secret it should be mounted into a container's
/etc/ssl/certs path. This way every https client will have the CA cert in
the trusted pool.

Signed-off-by: Roy Golan <rgolan@redhat.com>