script: Printing time in console.
Adds a new feature which prints a summary of the time it took in
each stage of the installer after the stage has been completed.

Summary is currently supported for the create, wait-for and the
delete command. The summary is printed at the completion of the
command and will show the time elapsed for each of the stages
of the execution.

Create command has the following stages:-
. Infrastructure
. API
. Bootstrap Complete
. Destroy bootstrap complete.
. Cluster Operator Complete.

Destroy command has the following stages:-
. Cluster
. Bootstrap

The wait-for command has the following stages:-
- Bootstrap-complete command
	. Bootstrap
	. API
- Install-complete command
	. Cluster
	. Console