pkg/types/aws: add service endpoints to aws platform
Adds a list of service endpoints to the AWS platform. A service endpoint entry/item consists of the service name and the endpoint url.

only one service endpoint for a service can be provided, i.e duplicates are not allowed.
the endpoint URL must be a hostname or a http(s) URL with no request path or requests queries.