gcp upi: enable internal load balancers
This change adds 02_lb_int.py template to the workflow to enable
internal load balancers. The cluster will begin communicating to the api
and mcs through the internal load balancers. The external load balancer
can optionally be disabled for private clusters.

This change also updates the documentation to use the $(command) syntax
to be in line with the other platforms.

In addition, the variable definitions were all moved to immediately
after the associated resources were created. This will help make clear
where their origins are.