asset/installconfig/aws/session.go: allow creating session with region and service overrides
Adds `GetSessionWithOptions` which allows creating aws.Session with options. Currently adds two options,

`WithRegion`, this configures the aws.Session to a specific region.

`WithServiceEndpoints`, this configures the aws.Session with a new service resolvers that uses the provided list of service endpoints to override the endpoints hard coded [1] into the AWS SDK for the specific service.
It uses an internal `awsResolver` that can be configures with a list of service endpoints such that when AWS SDK tries to fetch the endpoint for a service in a region, it returns the configured endpoint. If no such endpoint was configured for this internal resolver, it falls back to AWS SDK's default resolver that uses the hard coded list [1].
NOTE: the internal resolver returns the same endpoint URL for all regions for a service. This is acceptable as service endpoints provided to the installer don't provide region. Also since clusters are installed to single region, one should only need the endpoint for that region or global.

The aws metadata now provides consumers session initialized with region and AWS service overrides set in the install config.
The clients in various assets therefore transparently use the correct endpoints without extra configuration, while also allowing them to override it if needed when initializing the client.

[1]: https://github.com/aws/aws-sdk-go/blob/v1.29.31/aws/endpoints/defaults.go#L94