aws: configure terraform with service endpoints
The installer now passed on the service endpoints set in the install-config as a map of serivce name => service endpoint URL

- Why the map lookup when setting endpoints for aws provider

The endpoints for aws provider are a `schema.Set` [1] type and therefore an input map object cannot be directly assigned to the endpoints and therefore, we lookup the map object for specific key and set the corresponding endpoint.

There are two major concerns with this,

First, what if we need multiple endpoints for a service, like per region?
Terraform only supports providing ONE sevice endpoint per service, and this is fine as installer uses the terraform to create resources only in one region today, if there was a case where we had to create resources across regions, we might hav to extend the installer to invoke terraform multiple times.. there is no plan/requirement for that anytime soon.

Secondly, why is the list hard coded? that's not good right?
Installer only creates a SPECIFIC set of resources using the terraform, and therefore the hard coded list should be fine and it can be extended when we need new types of resources. This list is hard coded today because of terraform's limitations today.

[1]: https://github.com/terraform-providers/terraform-provider-aws/blob/81c55de44d67ad4515e615b9a5992a65be06bff6/aws/provider.go#L1236-L1240