destroy/aws: use the service overrides for AWS apis
Currently the destroy code-path for the isntaller is driven completely based on the `metadata.json`. This allows users to clean up the clusters as long as they possess/create a `metadata.json` for the cluster. Therefore, the service endpoints for the cluster need to be propagated to the `metadata.json` for later use.

Adds a list of service endpoints similar to the install-config to the metadata.json's definition. The metadata assets uses the install-config to set the service endpoints when provided by the user.

Updates the destroy/aws.New to create session using the `GetSessionWithOptions` similar to other assets to re-use the regions and service endpoints options. This makes the destroy code a lot less complicated.

`destroy/aws` ClusterInstaller allows users to provide no aws session. In this case, destroy continues to create aws session using SDK defaults with specified region and no service endpoint overrides. Requiring the user to handle special session and passing it along seems sane for users using destroy code directly similar to installer's metadata.json powered code-path.