platformtests: drop aws as no longer required
the only test defined are `TestGetDefaultInstanceClass`. That test are useful for developers of installer to make sure the default instance class used for a region is acurate and available in that region.

But these tests are no longer required since the installer moved to use AWS APIs to find out the correct default instance type [1]

[1]: https://github.com/openshift/installer/commit/d45e881abe1e3033b75df71fb537caa40f2834d8