aws: use SDK and rhcos AMIs to calculate known regions
w.r.t to the installer there are four kind of regions that user can provide.

First, is a commercial AWS region where the RHCOS AMIs are published and known to the installer
Second, is a commercial AWS region where the RHCOS AMIS are not published or not know to the isntaller at that verion
Third, is a restricted or soverign AWS region where RHCOS AMIs are not published but also these regions do not allowing traffic across the region traffic.
Fourth lastly is a custom region only available to the user.

For commercial regions in first, the installer is expected to use the AMI embedded in the binary. For the commercial regions in second, the installer is expected to `import` the AMI from one of the regions in first. For the restricted AWS regions, the installer cannot `import` the AMI to these regions and therefore the user is expected to provide the installer with the AMI that needs to be used. Similarly for custom regions only for users, the users are expected to provide the AMI.

Now, the list of regions that fall into first vs second depends on the regions for which AMI exists in `rhcos{-amd64}.json`. Since the contents of that file are mounted into the binary at build time using `shurcooL/vfsgen` there are two ways to provide this list to the go code,

- either provide an API as part of `rhcos` pkg that reads the rhcos.json from the embedded assets
since the file asset is not available to `read` unless the binary is built or `OPENSHIFT_INSTALL_DATA` env is set, if we were to run unit tests using `go test` the tests would fail to run. Having the capability of running unit tests without setting any env or building a release binary severly hinders the extent of the unit tests that are capable.

- check in a hard coded list of the regions based on the rhcos.json contents.
This has the advantage of go tests just working fine without any extra configuration. But now there is a need to update the file to match the rhcos.json
To help with automatic management, we can use go generate to create a go file that uses the rhcos.json as input and outputs a go file with a list of regions as a variable. But this file needs to be checked in the version control.

and the only step when the rhcos.json is bumped or changed is running `go generate ./pkg/rhcos/...`

To follow the second method, a new executable go file `ami_regions_generate.go` is added which can be executed as
```
go run ami_regions_generate.go <packae of the generated file> <source rhcos.json path> <destination go file path>
```
The generator provides a variable `AMIRegions` which is a string slice of all the regions that are part of `rhcos.json`

The `ami.go` include a go generate tag to allow go generate to execute the generator and update the list.

The installconfig asset package now provides a public function to check whether a region is `Known` i.e. regions in the first category.