asset/installconfig/aws: add validation to require endpoints for certain services for custom regions
Since the AWS SDK has the endpoints for most of the services for it's internal region, we require that users provide service endpoints for some required services when using custom region.

the required services are: "ec2", "elasticloadbalancing", "iam", "route53", "s3", "sts", and "tagging".

Also the users need to specify AMI for these custom regions.