aws: copy AMI from us-east-1 for SDK regions when not available
for the second category of the regions, commercial AWS regions that don't have RHCOS AMI published but that allow AMI import, the installer is expected to copy the AMI to the user's account from one of the RHCOS published ones.

So the installer picks the AMI published for us-east-1 region and copies it to the chosen region for the cluster to use.

Currently, the rhcos.Image asset's string content is extended to be `<ami-id>,<ami-id-region>` when the AMI to be used is different from the cluster's region. And the machinesets and machine objects are updated to use AWS tags to find the generated AMI.

The user is still allowed to provide an AMI if they want to re-use one, and the installer will use the AMI as is and will not copy/import the AMI.