images/baremetal: remove nss package as it's no longer needed
These images are failing to build internally for OSE due to missing the
nss wrapper package.

The underyling problem we are seeing when trying to use the
baremetal-installer image in CI to ssh to the packet server is that
user ID's are generated dynamically. This results in errors like this
when attempting to use SSH:

        No user exists for uid 102400000

The way most CI containers work around this is to echo a valid user to
/etc/passwd, as /etc/passwd is left writeable. The base images for
OpenShift now do this by default, so we can remove nss entirely.