images/baremetal: make /etc/passwd writeable
We'll be using baremetal-installer as the container we setup packet
for metal IPI CI. We need to use SSH, but it's broken without having a
valid user in /etc/passwd.

This change is required for our CI infrastructure to use SSH correctly
until we have completely moved CI to 4.x infrastructure where
/etc/passwd becomes dynamic.

This was restored to the tests container[1], and is now also being added
to the baremetal/installer container.

[1] https://github.com/openshift/origin/pull/24690

(cherry picked from commit aba60dad9372d1ea657f2dede6a76a93e2879eb8)