pkg/types/azure: add OutboundType for controlling egress
OutboundType is a strategy for how egress from cluster is achieved

Currently 2 strategies are:

- loadbalancer (default)
this uses the standard load balancer pointing to the control plane and compute nodes to provide egress based on [1]
for internal cluster, this means an public lb will be used to provide egress, i.e. even though all the endpoints of the cluster are internal and only accessible to the virtual network, there exists a public lb with public ip that provides the egress for the cluster

- user defined routing
this allows the user to configure the routing of the virtual network on their choosing. the installer is not expected to setup any egress, rather the installer should turn off egress from the standard lb. users have an option to setup the egress using [2]. An example is how AKS docuements using UDR for egress [3].
This can also be used if the user is using all egress to internet using a proxy, or when the user expects no egress to the internet at all.
Since the routing for the virtual network needs to setup by the user up-front before installing a cluster, using this strategy would require a pre-existing network

[1]: https://docs.microsoft.com/en-us/azure/load-balancer/load-balancer-outbound-connections#lb
[2]: https://docs.microsoft.com/en-us/azure/virtual-network/virtual-networks-udr-overview
[3]: https://docs.microsoft.com/en-us/azure/aks/egress-outboundtype