manifests/azure: no outbound k8s service type LB for UDR
For internal clusters, a dummy k8s service type LB is created for cluster. This service creates a public standard LB which is then used by compute nodes as an egress pathway to internet. Without this compute nodes in internal cluster would have no way to reach out to the internet.

But when the user is not using LB outbound type, i.e UDR, this service is no longer required because the user already has setup the network.