Option to set subnet and create Cluster SNAT
Add field to input file like:
platform:
  openstack:
    aciNetExt:
      clusterSNATSubnet: 10.0.0.1/16

The clusterSNATSubnet has to be a valid subnet or IP address. In the latter case, our CNI treats it as a /32 subnet. If this field is not provided, the installer wont create a yaml manifest for this CR.

The input above will create inside manifests directory a yaml called "cluster-network-27-snat-policy-cr.yaml":

apiVersion: aci.snat/v1
kind: SnatPolicy
metadata:
  name: clusternetworksnatpolicy
spec:
  snatIp:
    -  10.0.0.1/16

- Also minor changes to error conditions in acc-provision validation to sync with the exisiting structure