Bug 1815133: osp UPI machineset OS image name
The UPI documentation recommends keeping the worker Machineset in case
the user wants to create compute machines via the machine API. However,
the Machineset won't work unless the `image` property is updated to the
user-defined value.

This change adds a recommendation to update the OS image in case the
user wants to use the installer-provisioned Machineset.