images/baremetal: save rhcos.json in container image
For the baremetal platform, we need to know where the images are in
order to mirror them.  Especially for IPv6 environments where we may not
have outbound internet access to download the images.

Typically, we download the rhcos.json from the sha reported by
`openshift-install version` but this doesn't work in CI, where PR's are
always rebased on master and the sha reported doesn't exist on GitHub.

This is a temporary workaround until the machine OS content is part of
the release image itself.