Bug 1812950: generate metadata before tfvars
Now we generate terraform config and create resources for OpenStack
first, and only then we create the metadata.json file.
In case the resources were not created because of an error, we get
garbage in the system. And the installer cannot remove it because the
metadata file has not been generated yet.

This commit creates the file before the generation of terraform config.