ovirt: extend ovirt's MachinePool
Those items are now part of MachinePool
- instance type
- cpu (cores, sockets)
- memory
- os disk size
- vm type

With those we have controll on the way create worker/nodes,
and we have proper defaults set in the machine pool definitions, one for
the pkg/asset/machines/ master.go and worker.go.

Default master pool:
  cpus: 4
  mem: 16 GiB
  os disk: 120 GB
  VM type: high_performance

Default worker pool:
  cpus: 4
  mem: 16 GiB
  os disk: 120 GB
  VM type: server

A compute pool snippet from the install-config.yaml may look like that:

```yaml
compute:
- architecture: amd64
  hyperthreading: Enabled
  name: worker
  platform:
    ovirt:
      cpu:
        cores: 8
        sockets: 1
      instanceTypeID:
      memoryMB: 65536
      osDisk:
        sizeGB: 120
      vmType: high_performance
  replicas: 3
```

Terraform now uses those values, and all the defaults in tf files are
now deleted in favour of the machine-config ones.
To support `osDisk` size which is differnt than the template
master/main.tf has this block:

  block_device {
    interface = "virtio_scsi"
    size      = var.ovirt_master_os_disk_size_gb
  }

If needed TF will extend the disk size of the VM.

Cluster-api-provider-ovirt also extends the disk of a VM if the
definition in the machine spec is differnt than the template.

Signed-off-by: Roy Golan <rgolan@redhat.com>