vSphere: Add IPI-specific platform validation.
This commit creates a new asset to perform platform validation specifically for cluster creation. This is useful for when UPI and IPI install configs may have different requirements. In the case of vSphere, IPI requires cluster and VIPs but these are optional in UPI.

IPI requires the following fields which are optional for UPI:
- Network
- Cluster
- APIVIP
- IngressVIP
- DNSVIP