Validate Azure regions
Regions were already being filtered by two rules: the regions available for the subscription in context,
and the ones that support the creation of resources by the given subscription. That filtering however
happened only at the survey CLI level, meaning invalid regions would not be displayed as an option, but would
be accepted if the users provided, for example, their own install-config file specifying an invalid region,
resulting in later stages of the resources creation.

This change applies the same filtering to actual validations and adds proper error messages in case invalid
regions are provided.