Baremetal: Bug 1801238: Pull data from ironic inspector and annotate BareMetalHost
Right now we don't get any introspection data about the masters
because they are provisioned by the bootstrap node and the information
is lost.  This patch adds a script to pull the introspection data
from ironic running on the bootstrap and places it in the BareMetalHost CRs
for the masters.  This is done via an annotation as we cannot directly
update the Status in BMH.  The annotation is picked up by the BMO
and added to the BMH CRs.

There are several bugs filed about this:

https://github.com/openshift-metal3/dev-scripts/issues/917
https://bugzilla.redhat.com/show_bug.cgi?id=1801238