always atleast use the default SSH key for gathering
The gather command will always add the bootstrap-ssh keypair to the SSH client when trying to gather the logs from bootstrap-host.

The bootstrap-ssh keypair is NOT used when the user alrady has a SSH_AGENT set, modifying the user SSH_AGENT to add a bootstrap host only key pair would not be a sane decision.