data/data/azure: Unify cloud LB and kube-apiserver LB
Azure only allows a NIC to be in one public and one internal load
balancer, which prevents service load balancers from working on
masters (the NIC is part of the kube-apiserver LB). This prevents
compact Azure clusters and any future service that wishes to run
on the master.

Instead, rename the Azure LB and backend pool so that it matches
the names the Azure cloud controller will add service load balancer
rules to. Azure LBs allow multiple frontends against a single
backend and so both kube-apiserver (selecting just masters) can
co-exist with service load balancers. Both public and internal
pools are selected.

The only downside is a pod on a node on port 6443 would be included
in the balancer - but since we intend to move kube-apiserver behind
a service load balancer in the future and nothing stops the router
from being impacted the same way (someone getting root on any node
can add service load balancer frontends), it is no worse than current
behavior. Note that these injected endpoints would still require a
valid HTTPS certificate and so could not actually impersonate the
master.

Compact clusters and service load balancers spanning masters and
workers should now function.