Bug-1802526: update AWS TF to use all IPv4 machine CIDRs
This commit updates the AWS platform in Terraform to use all IPv4 machine CIDRs,
instead of just the first. This achieves:
- further progress on deprecating use the individual machine CIDR var.machine_cidr
- resolving BZ-1802526 by adding all CIDRs to the security group

BZ-1802526 shows that in BYO VPC & subnet scenarios the VPC may have multiple cidr
associations and the subnets do not necessarily belong to the main cidr block. This
lead to failures when the only the main VPC CIDR block was added to the security
group, but the subnets were in a different valid block.