pkg/destroy/ovirt: Cleanup VM logging
Fixing some of the logging from 05177839ab (ovirt: Implement destroy,
2019-10-28, #1948) to avoid things like:

  $ openshift-install destroy cluster --dir one
  INFO searching VMs by tag=one-6xlb7
  INFO Found %!s(int=3) VMs
  INFO Stopping VM one-6xlb7-master-1 : errors: %s%!(EXTRA <nil>)
  ...

because the final 'err' had no %s in the template string to consume
it.  This commit fixes the two cases which had an 'errors: %s' using
up the final template %s, but leaves a lot of other potential logging
cleanup in this package alone.