aws: add missing endpoints for aws china route53
since, AWS SDK is missing the route53 endpoints for AWS China regions, the users will be forced to add the service endpoint.
But it seems like there is a known endpoint for that service already and the sdk is trailing for a bit, also the terraform aws provider is using this same endpoint already.
So adding the endpoint as default known, should help our users a little bit...

Also the destroy code needs to perform the same switching for AWS china partition like we go for AWS commercial regions for resource tagging api wrt. route53 object from [1]

[1]: https://github.com/openshift/installer/commit/e24c7dc419706b67f9ee2c0bc21f01ebeafc4833