installconfig/ovirt: fix CA prompt and add warning
This is a follow-on to 3eccc348. After looking at this more closely, I
realized that the cert bundle is only prompted if the CA is trusted
locally. This means that if the CA isn't trusted, the communication with
the oVirt engine will be insecure. I added a warning to reflect that
reality.