GCP: Fixed error message for Invalid Project ID entry
Fixes the error message that occurs after the region selection when an
invalid project ID is entered. Solved by picking up the list of projects
that the user has in their GCP account and offers a select option on the
project entry to select the one that they want. Reduces the chance of
picking an invalid project ID and avoids the error.

CloudResourceManager is used to pick up the list of projects for current
user and survey is populated with the values obtained.

Generated the gcpclient_generated.go mock file using mockgen and replaced
the existing file to add the GetListOfProjects function. Running
hack/go-genmock.sh did not work after I changed the interface and ran it.