Change local_only=true to local_only=false  #1007
libvirt: Unable to access web console #1007

If the optional localOnly attribute on the domain element is "yes", then DNS requests under this domain will only be resolved by the virtual network's own DNS server - they will not be forwarded to the host's upstream DNS server. If localOnly is "no", and by default, unresolved requests will be forwarded.

Which will resolve the issue of authentication and console pods of not resolving the dns names.