baremetal: update provisioning CR to quote strings
For the provisioning config, it's allowable for provisioningDHCPRange to
be empty. If it is, the current template produces this YAML:

```yaml
spec:
   provisioningDHCPRange:
```

Unfortunately, in YAML-land the above is null not empty string, so when
you apply it, you get complaints:

```
Provisioning.metal3.io "provisioning-configuration" is invalid:
spec.provisioningDHCPRange: Invalid value: "null":
spec.provisioningDHCPRange in body must be of type string: "null"
```

This change ensures all string values are quoted, which will result in a
correct value for provisioningDHCPRange when empty.