pkg/asset/machines/vsphere: set default folder value for Machine API
Sets the Machine API to use the folder value if specified in the install-config. Otherwise defaults the folder value to the cluster id. Updates cloud provider config to use the cluster ID for default value, where it had been using the cluster name before.

In the machine pool we use the absolute path for the default cluster ID, e.g.: /<datacenter>/vm/<clusterID>
Although a relative path (that is, just specifying the folder name) seems to work for a top-level VM folder, I am using the absolute path to be consistent with the nested folder use case which requires an absolute path.

The cloud provider config, on the other hand seems to prefer a relative path based on the documentation: https://vmware.github.io/vsphere-storage-for-kubernetes/documentation/existing.html