openstack UPI: Server group name to match IPI
The installer creates a server group as a side effect of the manifest
generation.

Prior to this patch, the UPI playbooks created yet another server group
and left the installer-created one unused.

With this patch, the UPI playbooks reuse the server group created by the
installer, rather than creating an additional one.