Azure: Add functionality to change Azure Machine Disk Types
Azure disk types for the master and worker nodes is by default
set to Premium SSD(Premium_LRS) setting. Made the changes to add
the disktype of the master and worker nodes in the install config
by adding an extra field in MachinePool DiskType to the OSDisk struct
which is part of the DefaultMachinePlatform attribute in the Azure
Platform type. To change the disk type, the install config must be
changed by adding the diskType under platform->Azure property by adding
the defaultMachinePlatform->osDisk->diskType and setting the value to
one of the following, Premium_LRS and StandardSSD_LRS.

For custom Azure modifications to the disk types, the worker nodes can have
an additional disk type Standard_LRS. HDD type disks for master machines are
not recommended so it cannot be added to the default values or to the
custom modifications for the master.

Currently, the UltraSSD_LRS is under Azure preview and it is only
available in three regions and hence is currently disabled(commented).