pkg: add explain package
The explain package provides an `explain` cobra command that can use used to add it as subcommand.

The package includes 2 major functions,

- fields_lookup
Given a list of paths, this finds the schema for the path from the install config schema.

- printer
The printer printers the schema provided to the function

The printer always prints the InstallConfig apiVersion, kind to the provided io.Writer
It prints the description of the field using the schema provided,
and for object type, it also prints all the properties of the object and their descriptions.