data/data/aws/vpc: create subnet sizes based on az count
For every availabilty zone, we need a subnet. `ceil(log2(num of AZs))` finds the smallest exponent of 2 that is greater than or equal to `num of AZs`

Using that to split the private/public machine_cidr range allows the installer to carve out network ranges for each AZ with minimum unused.
Previously the installer would create 8 ranges always even when only a few were required.