Bug 1836042: set folder absolute path in vSphere cloud provider
This is a partial revision of 814f64f4bce13a1c4472a104f72b7658b0e76abe,
where we update to use an absolute path rather than a relative path.
Using a relative path results in an error when trying to provision a PVC
and a subfolder was used for install.