gather: enable bootstrap log gathering for baremetal
This enables bootstrap log gathering for baremetal. We fetch the IP's
for the control plane hosts using ironic introspection data, and pick
the best one. We prefer to find an IP in a machine network, but
otherwise we'll use the first valid IP found. For the bootstrap, the
only IP address we can know is the provisioning IP for the bootstrap.