vendor: replace terraform-provider-vsphere with OpenShift fork
With https://bugzilla.redhat.com/show_bug.cgi?id=1825323 we see terraform
failing with an inconsistent plan when the vSAN storage policy enforces
thin-provisioned disks but for some reason the terraform provider produces
a plan stating they are thick-provisioned. Terraform docs indicates these
types of problems require a fix to CustomizeDiff in the provider:
https://www.terraform.io/docs/extend/terraform-0.12-compatibility.html#inaccurate-plans

In a recent PR
https://github.com/terraform-providers/terraform-provider-vsphere/pull/1052
we see that when the VM is being imported from an OVA, the DiskDiffOperation
is simply skipped. DiskDiffOperation is the function that produces the
inconsistent plan in the bugzilla mentioned before.  We have created a
similar PR to skip DiskDiffOperation when VMs are being cloned:
https://github.com/terraform-providers/terraform-provider-vsphere/pull/1075

This commit carries that patch until it has merged upstream or resolved otherwise

The commit was created by running:
 go mod edit -replace github.com/terraform-providers/terraform-provider-vsphere=github.com/openshift/terraform-provider-vsphere@v1.18.1-openshift-1
 go mod edit -replace github.com/vmware/govmomi=github.com/vmware/govmomi@v0.22.2-0.20200420222347-5fceac570f29
 go mod tidy
 go mod vendor

This commit also updates the vsphereprivate provider to use the updated libraries.