vendor: use openshift fork of terraform-provider-aws and bump to v2.60.0
This is to address the "NoSuchBucket: The specified bucket does not exist"
error as explained in https://bugzilla.redhat.com/show_bug.cgi?id=1759617 and
many other similar bugs. This bug has been "fixed" several times over the
years, yet it continues to rear its ugly self. The ultimate problem is a race
condition with S3 eventual consistency. As described in the bug above, the
bucket does not yet exist when trying to reference tags. The openshift fork
that this commit references, contains an upstream patch as described in
https://github.com/terraform-providers/terraform-provider-aws/pull/13009 that
should address this issue.