support for other Azure cloud environments
Add the `azure.cloudName` field to the installconfig. This field
directs which Azure cloud environment is used for the cluster.

The `status.platformStatus.azure.cloudName` field is set in the
infrastructure.config.openshift.io resource to match the cloud name
configured in the installconfig.

The `cloud` field is set in the cloud-provider-config ConfigMap
to match the cloud name configured in the installconfig.

See https://github.com/openshift/enhancements/pull/321

https://issues.redhat.com/browse/CORS-1442