GCP: Set validations for GCP disk sizes.
Current disk size restrictions for GCP is set to be above 0.
The current size of the image is 16GB and hence the disk sizes
must at least be 16GB. Also, the maximum disk size limit on GCP
is 65536GB so users must not be allowed to create disks above that
limit. Added the validations to the install config input for GCP.