Bug 1836098: OpenStack: check that port exists before creating a fip
To create a floating IP we use the identifier of the port created before.
Unfortunately, due to a race condition in OpenStack Neutron, we may get a
situation when the identifier is already received, but the port is not yet
available. In this case we get the NotFound error when creating the
floating ip.
For us this happens when we try to create API floating IP in this module.

To prevent this race condition we explicitly create a chain by setting
the `depends_on`instructions:
`api_port` -> `ingress_port` -> `masters` -> `api_fip`
This solution gives us enough time to make sure that `api_port` is
available when we need to create `api_fip`