wait-for install-complete: Fix re-generation of install config
If the cluster is created and the wait-for install-complete command
is called to check the status, the installer checks if the install
config exists to get the platform information to set the timeout for
baremetal clusters a little higher. Since the install config is
consumed during the cluster creation, the installer will then start
to re-generate the install config and ask the user to provide the
information again.

Modified the platform information gathering to only pick the information
up only if the information is available and if it does not exist, set the
timeout value to the maximum value required as default. This will
now avoid re-generation of the install config.